Add Logout handling
Add Logout logic to logout button.
Wrap App in a component so that it can be connected to redux.
Add auth state change listeners to wrapped component.
Add login and logout actions that can be called from the listener.
Reset redux store on logout.
Switch to SwitchNavigators, which unmount the screens the user has
navigated away from. This is necessary for preventing weird behavior
when resetting the redux store on logout.
Remove bug in AuthInput, which came to life because the component was
unmounted in the new navigator type, and setting the state of an
unmounted component causes memory leak.